tgo app conversions and initial tgo_module_stitchUnstitch test conversion.
