+++ isis/src/tgo/apps/tgocassis2isis/tgocassis2isis.h
@@ -1,0 +1,11 @@
+#ifndef tgocassis2isis_h
+#define tgocassis2isis_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void tgocassis2isis(UserInterface &ui);
+}
+
+#endif
