+++ isis/src/tgo/apps/tgocassisstitch/tgocassisstitch.h
@@ -1,0 +1,11 @@
+#ifndef tgocassisstitch_h
+#define tgocassisstitch_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void tgocassisstitch(UserInterface &ui);
+}
+
+#endif
