+++ isis/src/tgo/apps/tgocassisunstitch/tgocassisunstitch.h
@@ -1,0 +1,11 @@
+#ifndef tgocassisunstitch_h
+#define tgocassisunstitch_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void tgocassisunstitch(UserInterface &ui);
+}
+
+#endif
